Ajuste na estilizacao da pagina inicial, adicionei algumas formatacoes no card do produto e na secao de grupos, adicionei efetivamente a exclusao, edicao de produtos e exclusao de usuarios, ajustei a criacao dinamica dos elementos no dashboard, ja esta trazendo tambem as ultimas compras do usuario e agora falta tratar melhor a exibicao ao clicar no produto para remover inconsistencias no carrinho

+++ routes/web.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\ProdutoController;
 use App\Http\Controllers\GrupoController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\HomeController;
+use App\Http\Controllers\VendaController;
 use Illuminate\Support\Facades\Route;
 Route::post('/buscar', [HomeController::class, 'buscarProduto']); //rota para fetch
 Route::post('/limpar/carrinho', [CarrinhoController::class, 'limparCarrinho']);
@@ -15,13 +16,19 @@ Route::post('/pesquisar/usuarios', [ProfileController::class, 'pesquisarUsuarios
 Route::get('/', [HomeController::class, 'index'])->name('home');
 Route::get('/produtos', [ProdutoController::class, 'index'])->middleware(['auth', 'verified'])->name('produtos');
 Route::get('/usuarios', [ProfileController::class, 'index'])->middleware(['auth', 'verified'])->name('usuarios');
-Route::delete('/usuarios/{id}', [ProfileController::class, 'destroyUser'])->name('usuarios.destroy');
-Route::delete('/produtos/{id}', [ProdutoController::class, 'destroy'])->name('produtos.destroy');
+Route::delete('/usuarios/{id}', [ProfileController::class, 'destroyUser'])->middleware(['auth', 'verified'])->name('usuarios.destroy');
+Route::delete('/produtos/{id}', [ProdutoController::class, 'destroy'])->middleware(['auth', 'verified'])->name('produtos.destroy');
 Route::get('/produtos/{id}/edit', [ProdutoController::class, 'edit'])->name('produtos.edit');
 Route::put('/produtos/{id}', [ProdutoController::class, 'update'])->name('produtos.update');
-Route::get('/dashboard', function () {
+Route::post('/registrar/venda', [VendaController::class, 'registrarVenda']);
+Route::get('/dashboard', [VendaController::class, 'listarComprasCliente'])
+    ->middleware(['auth', 'verified'])
+    ->name('dashboard');
+
+
+/*Route::get('/dashboard', function () {
     return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+})->middleware(['auth', 'verified'])->name('dashboard'); */
 
 
 
@@ -32,4 +39,4 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-require __DIR__.'/auth.php';
+require __DIR__ . '/auth.php';
